BKRS-271 Replace os.Stat fallback with OS-specific isNotDir helper
