+++ io/storage/local/utils_unix.go
@@ -1,0 +1,7 @@
+//go:build !windows
+
+package local
+
+func isWindowsNotDir(err error) bool {
+	return false
+}
